FIR IDE: add TypeInfoProvider and mixin to host type utils
