+++ idea/idea-frontend-fir/src/org/jetbrains/kotlin/idea/asJava/firLightUtils.kt
@@ -28,6 +28,7 @@ import org.jetbrains.kotlin.fir.symbols.impl.FirClassSymbol
 import org.jetbrains.kotlin.fir.types.*
 import org.jetbrains.kotlin.idea.fir.low.level.api.api.FirModuleResolveState
 import org.jetbrains.kotlin.idea.fir.low.level.api.api.withFirDeclaration
+import org.jetbrains.kotlin.idea.frontend.api.components.DefaultTypeClassIds
 import org.jetbrains.kotlin.idea.frontend.api.fir.analyzeWithSymbolAsContext
 import org.jetbrains.kotlin.idea.frontend.api.fir.symbols.KtFirSymbol
 import org.jetbrains.kotlin.idea.frontend.api.fir.types.KtFirType
@@ -35,6 +36,7 @@ import org.jetbrains.kotlin.idea.frontend.api.symbols.*
 import org.jetbrains.kotlin.idea.frontend.api.symbols.markers.*
 import org.jetbrains.kotlin.idea.frontend.api.types.*
 import org.jetbrains.kotlin.load.kotlin.TypeMappingMode
+import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.name.SpecialNames
 import org.jetbrains.kotlin.types.model.SimpleTypeMarker
 import java.text.StringCharacterIterator
@@ -305,6 +307,12 @@ internal fun KtType.getTypeNullability(context: KtSymbol, phase: FirResolvePhase
     return if (isNotPrimitiveType) NullabilityType.NotNull else NullabilityType.Unknown
 }
 
+internal val KtType.isUnit get() = isClassTypeWithClassId(DefaultTypeClassIds.UNIT)
+
+internal fun KtType.isClassTypeWithClassId(classId: ClassId): Boolean {
+    if (this !is KtClassType) return false
+    return this.classId == classId
+}
 
 private fun escapeString(str: String): String = buildString {
     str.forEach { char ->
